Delete grpcAPI
