+++ GroupProject_PRN231_NET1606_TRY_eJournal.sln
@@ -23,8 +23,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Infrastructure", "Infrastru
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "gRPCDataAccess", "gRPCDataAccess", "{DB199C2B-DE9E-4EF4-B640-AF98F12F3897}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "API_gRPC", "API_gRPC\API_gRPC.csproj", "{4A87FA65-41BD-4651-9243-6B6120CA569F}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GrpcService", "GrpcService\GrpcService.csproj", "{A0390706-F606-4478-A0E0-89CCD1C8C8DC}"
 EndProject
 Global
@@ -53,10 +51,6 @@ Global
 		{E9B0EB85-6F22-497B-A8AE-C8E937E8CB9A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E9B0EB85-6F22-497B-A8AE-C8E937E8CB9A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E9B0EB85-6F22-497B-A8AE-C8E937E8CB9A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4A87FA65-41BD-4651-9243-6B6120CA569F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4A87FA65-41BD-4651-9243-6B6120CA569F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4A87FA65-41BD-4651-9243-6B6120CA569F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4A87FA65-41BD-4651-9243-6B6120CA569F}.Release|Any CPU.Build.0 = Release|Any CPU
 		{A0390706-F606-4478-A0E0-89CCD1C8C8DC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A0390706-F606-4478-A0E0-89CCD1C8C8DC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A0390706-F606-4478-A0E0-89CCD1C8C8DC}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -71,7 +65,6 @@ Global
 		{6854D0AB-1213-41F3-842F-B491FFDFA7DF} = {D9F00161-EA78-4FED-B9C4-852BD08EFBB1}
 		{D0E17860-28EA-4235-8C5D-F92DAA658F2E} = {624F97ED-42B1-4A8F-9185-37D51B6DC4D3}
 		{E9B0EB85-6F22-497B-A8AE-C8E937E8CB9A} = {CE41E970-DA96-45E6-8C6A-C7CDE103C0FF}
-		{4A87FA65-41BD-4651-9243-6B6120CA569F} = {227DD01C-6E9D-49F8-8A49-AD4937773092}
 		{A0390706-F606-4478-A0E0-89CCD1C8C8DC} = {DB199C2B-DE9E-4EF4-B640-AF98F12F3897}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
