+++ GroupProject_PRN231_NET1606_TRY_eJournal/Protos/issueCRUD.proto
@@ -1,8 +1,7 @@
 syntax = "proto3";
 
-option csharp_namespace = "API.issueCRUD";
+option csharp_namespace = "Api.issueCRUD";
 package issue;
-import "google/protobuf/timestamp.proto";
 service IssueCRUD{
     rpc GetAllIssue(Empty) returns (ListIssues);
     rpc CreateIssue(AddIssue) returns (IssueStatus);
@@ -17,7 +16,7 @@ message ListIssues {
 message IssueViewModel{
      string Volumn =1;
        string  Description =2;
-    google.protobuf.Timestamp DateRelease =3;
+    string DateRelease =3;
   repeated ArticleViewModel Articles =4;
 }
 message ArticleViewModel {
