+++ GroupProject_PRN231_NET1606_TRY_eJournal/SchemaFilter/AddIssueSchemaFilter.cs
@@ -1,9 +1,9 @@
-﻿using API.issueCRUD;
+﻿using Api.issueCRUD;
 using Microsoft.OpenApi.Any;
 using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.SwaggerGen;
 
-namespace API_gRPC.SchemaFilter
+namespace GroupProject_PRN231_NET1606_TRY_eJournal.SchemaFilter
 {
     public class AddIssueSchemaFilter : ISchemaFilter
     {
