+++ GroupProject_PRN231_NET1606_TRY_eJournal/SchemaFilter/UpdateIssueSchemaFilter.cs
@@ -1,4 +1,4 @@
-﻿using API.issueCRUD;
+﻿using Api.issueCRUD;
 using Microsoft.OpenApi.Any;
 using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.SwaggerGen;
