+++ GrpcService/Repository/IssueRepository.cs
@@ -1,5 +1,5 @@
 ﻿using BusinessObject;
-using GrpcService.InterfaceRepository;
+using Application.InterfaceService;
 using GrpcService.InterfaceService;
 using Microsoft.EntityFrameworkCore;
 using System;
@@ -7,7 +7,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-
+using GrpcService.InterfaceRepository;
 namespace GrpcService.Repository
 {
     public class IssueRepository : GenericRepository<Issue>, IIssueRepository
